+++ CHANGELOG.md
@@ -3,11 +3,18 @@ All notable changes to this project will be documented in this file.
 <br />The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 <br />and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.1.0] - 2020-07-09
+## [3.1.0] - 2020-07-10
 ### Added
 - New command: `!cast` – Rolls a spell's Power Level.
 - New command: `!mishap` – Draws a random Magic Mishap.
-- Swedish language support for MYZ and FBL crit tables.
+- New command: `!mutation` – Rolls dice for a Mutation and checks for any Misfire.
+- New command: `!feral` – Rolls dice for a GenLab Alpha Animal Power and checks for any Feral Effect.
+- New command: `!module` – Rolls dice for a Mechatron Module and checks for any Overheating.
+- New command: `!contact` – Rolls dice for an Elysium Contact and checks for any Backlash.
+- New command: `!myzpower` – Rolls dice for a MYZ power.
+
+### Changed
+- Renamed the old `!mutation` command into `!drawmutation`.
 
 ### Fixed
 - Typos in both Readme files.